requirements.txt
python-docx
python-pptx
openpyxl
gradio
influxdb_client 
scikit-learn
openai
PyPDF2
importnb
plotly
python-dotenv